Update test cases for double

+++ src/test/java/seedu/address/model/person/PersonTest.java
@@ -104,13 +104,13 @@ public class PersonTest {
     public void updateBalance() {
         // update balance by adding
         Person person1 = new PersonBuilder().build();
-        int initialBalance1 = person1.getBalance();
+        double initialBalance1 = person1.getBalance();
         person1.updateBalance(20);
         assertEquals(initialBalance1 + 20, person1.getBalance());
 
         // update balance by deducting
         Person person2 = new PersonBuilder().build();
-        int initialBalance2 = person2.getBalance();
+        double initialBalance2 = person2.getBalance();
         person2.updateBalance(-20);
         assertEquals(initialBalance2 - 20, person2.getBalance());
     }
